backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.3"
dependencies = [
    "alembic==1.17.2",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "bcrypt==4.3.0",
    "cached-property==2.0.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "cryptography==46.0.3",
    "deprecation==2.1.0",
    "ecdsa==0.19.1",
    "fastapi==0.104.1",
    "gotrue==1.3.1",
    "greenlet==3.2.4",
    "h11==0.14.0",
    "httpcore==0.17.3",
    "httpx==0.24.1",
    "idna==3.11",
    "inflection==0.5.1",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "postgrest==0.13.2",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "realtime==1.0.6",
    "requests==2.32.5",
    "resend==2.21.0",
    "rsa==4.9.1",
    "sentry-sdk==2.48.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "starlette==0.27.0",
    "storage3==0.6.1",
    "strenum==0.4.15",
    "stripe==14.0.1",
    "supabase==2.0.2",
    "supafunc==0.3.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "urllib3==2.5.0",
    "uvicorn==0.24.0",
    "websockets==12.0",
]

[dependency-groups]
dev = [
    "black==26.1.0",
    "pyright==1.1.398",
]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
exclude = [
    "**/__pycache__",
    ".venv",
    "venv",
]

[tool.black]
line-length = 88
target-version = ["py312"]
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
)/
'''
